+++ controllers/webhookcerts/certs.go
@@ -1,4 +1,4 @@
-package bootstrapper
+package webhookcerts
 
 import (
 	"crypto/rand"
